MA-1313/Renaming-package
- Renamed from "com.reactlibrary" -> "com.reactlibrary.ymchat"
