+++ android/src/main/AndroidManifest.xml
@@ -1,6 +1,4 @@
-
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.reactlibrary">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.reactlibrary.ymchat">
 
 </manifest>
