+++ android/src/main/java/com/reactlibrary/ymchat/YMChatModule.java
@@ -1,5 +1,4 @@
-
-package com.reactlibrary;
+package com.reactlibrary.ymchat;
 
 import com.facebook.react.bridge.Callback;
 import com.facebook.react.bridge.ReactApplicationContext;
